% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fastlm.R
\name{summary.fastlm}
\alias{summary.fastlm}
\title{summary.fastlm}
\usage{
\method{summary}{fastlm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fastlm}.}
}
\value{
Prints the summary to the console.
}
\description{
Provides a summary for objects of class \code{fastlm}.
}
\examples{
set.seed(123)
n <- 1000
p <- 10
X <- matrix(rnorm(n * p), ncol = p)
beta_true <- rnorm(p + 1)
y <- cbind(1, X) \%*\% beta_true + rnorm(n)
model_fast <- fast_lm(X, y)
summary.fastlm(model)
}
